mt_evtMain.c: FUN_0036d030 retype param_1 long->int (nd588->529, obj832->808)

+++ src/Event/mt_evtMain.c
@@ -619,7 +619,7 @@ void FUN_003663f0(u32 param_1,int param_2);
 void FUN_00366660(u32 param_1,u64 param_2,u64 param_3,u64 param_4);
 void FUN_00366970(u64 param_1,u64 param_2);
 void FUN_00366ac0(u32 param_1,u64 param_2);
-void FUN_0036d030(long param_1);
+void FUN_0036d030(int param_1);
 void FUN_0036efb0(void);
 void FUN_0036f000(long param_1);
 void FUN_0036f260(void);
@@ -13379,7 +13379,7 @@ u64 FUN_0036cc30(u64 param_1,u64 param_2,u8 *param_3)
 // FUN_0036D030 NONMATCHING
 
 
-void FUN_0036d030(long param_1)
+void FUN_0036d030(int param_1)
